# openapi/mlbb_image_matching_engine_openapi_v1_3_1_icon_center_anchor.yaml
openapi: 3.1.0
info:
  title: MLBB Image Matching Engine V1.3.1
  version: v1.3.1-icon-center-anchor-20260705
  description: Icon-center anchored MLBB result-screen reference matching API. Candidate-only;
    no Hero Lock.
servers:
- url: https://mlbb-image-matching-engines.onrender.com
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Health check
      responses:
        '200':
          description: Health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /v1/mlbb/image-match/analyze:
    post:
      operationId: analyzeMLBBResultScreen
      summary: Analyze MLBB result screen
      security:
      - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
      responses:
        '200':
          description: Candidate-only matching report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeResponse'
        '400':
          description: Bad request
        '401':
          description: Invalid or missing API key
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
        engine:
          type: string
        version:
          type: string
        profile_id:
          type: string
        reference_icon_count:
          type: integer
        unique_hero_count:
          type: integer
      required:
      - status
      - engine
      - version
      - profile_id
      - reference_icon_count
      - unique_hero_count
    AnalyzeRequest:
      type: object
      properties:
        image_base64:
          type: string
        screen_type:
          type: string
          enum:
          - result_screen
          default: result_screen
        top_k:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
        return_debug_samples:
          type: boolean
          default: false
      required:
      - image_base64
    AnalyzeResponse:
      type: object
      properties:
        engine:
          type: string
        version:
          type: string
        profile_id:
          type: string
        screen_type:
          type: string
        hero_lock_allowed:
          type: boolean
        rule:
          type: string
        reference_icon_count:
          type: integer
        unique_hero_count:
          type: integer
        slots:
          type: array
          items:
            type: object
      required:
      - engine
      - version
      - profile_id
      - screen_type
      - hero_lock_allowed
      - rule
      - reference_icon_count
      - unique_hero_count
      - slots